James lowther

+++ src/sengproject/gui/researcher/tvobjects/ResearcherReviewer.java
@@ -3,12 +3,12 @@ package sengproject.gui.researcher.tvobjects;
 public class ResearcherReviewer {
 
     private String name;
-    private int num_major_rev;
-    private int num_minor_rev;
-    private int num_pub_rev;
-    private float avg_rev_time;
+    private long num_major_rev;
+    private long num_minor_rev;
+    private long num_pub_rev;
+    private double avg_rev_time;
 
-    public ResearcherReviewer (String n, int maj_r, int min_r, int pub_r, float avg_t) {
+    public ResearcherReviewer (String n, long maj_r, long min_r, long pub_r, double avg_t) {
 
         name = n;
         num_major_rev = maj_r;
@@ -22,20 +22,20 @@ public class ResearcherReviewer {
 
     public String getName () {return name;}
 
-    public void setNum_major_rev (int n) {num_major_rev = n;}
+    public void setNum_major_rev (long n) {num_major_rev = n;}
 
-    public int getNum_major_rev () {return num_major_rev;}
+    public long getNum_major_rev () {return num_major_rev;}
 
-    public void setNum_minor_rev (int n) {num_minor_rev = n;}
+    public void setNum_minor_rev (long n) {num_minor_rev = n;}
 
-    public int getNum_minor_rev () {return num_minor_rev;}
+    public long getNum_minor_rev () {return num_minor_rev;}
 
-    public void setNum_pub_rev (int n) {num_pub_rev = n;}
+    public void setNum_pub_rev (long n) {num_pub_rev = n;}
 
-    public int getNum_pub_rev () {return num_pub_rev;}
+    public long getNum_pub_rev () {return num_pub_rev;}
 
-    public void setAvg_rev_time (float n) {avg_rev_time = n;}
+    public void setAvg_rev_time (double n) {avg_rev_time = n;}
 
-    public float getAvg_rev_time () {return avg_rev_time;}
+    public double getAvg_rev_time () {return avg_rev_time;}
 
 }
